requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
openai>=1.50.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
streamlit>=1.40.0
pypdf>=5.0.0
python-multipart>=0.0.12
pandas>=2.0.0
tiktoken>=0.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
